requirements.txt
streamlit
openai
python-dotenv
langchain-community
bs4
azure.ai.inference
